+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.9.0]
 
+### Changed
+
+- Shapes panel (left sidebar) is now closed by default for a cleaner initial view (fixes [#484](https://github.com/hediet/vscode-drawio/issues/484))
+
 ### Fixed
 
 - Reverts change to automatically follow VS Code dark/light theme [#457](https://github.com/hediet/vscode-drawio/issues/457)